Extract global updater to separate class
