+++ contrib/ruby_event_store-persistent_projections/spec/global_updater_spec.rb
@@ -2,24 +2,25 @@ require 'spec_helper'
 
 module RubyEventStore
   module PersistentProjections
-    RSpec.describe Consumer, db: true do
+    RSpec.describe GlobalUpdater, db: true do
       include SchemaHelper
 
       let(:logger_output) { StringIO.new }
       let(:logger) { Logger.new(logger_output) }
+      let(:clock) { Time }
 
       specify "creates projection global status if doesn't exist" do
-        consumer = Consumer.new(SecureRandom.uuid, nil, logger: logger)
+        consumer = GlobalUpdater.new(logger: logger, clock: clock)
 
         consumer.one_loop
 
-        expect(ProjectionStatus.count).to eq(1)
+        expect(GlobalUpdater::ProjectionStatus.count).to eq(1)
         expect(global_status.position).to eq(0)
       end
 
       specify 'global thread progresses the state if available' do
         publish_event
-        consumer = Consumer.new(SecureRandom.uuid, nil, logger: logger)
+        consumer = GlobalUpdater.new(logger: logger, clock: clock)
 
         consumer.one_loop
 
@@ -28,7 +29,7 @@ module RubyEventStore
 
       specify 'global thread progresses the state if event locked' do
         publish_event
-        consumer = Consumer.new(SecureRandom.uuid, nil, logger: logger)
+        consumer = GlobalUpdater.new(logger: logger, clock: clock)
         expect(consumer).to receive(:check_event_on_position).with(1).and_raise(ActiveRecord::LockWaitTimeout)
 
         result = consumer.one_loop
@@ -40,7 +41,7 @@ module RubyEventStore
       specify 'global thread progresses the state if event not found' do
         publish_event_and_rollback
         publish_event
-        consumer = Consumer.new(SecureRandom.uuid, nil, logger: logger)
+        consumer = GlobalUpdater.new(logger: logger, clock: clock)
 
         consumer.one_loop
         result = consumer.one_loop
@@ -50,11 +51,11 @@ module RubyEventStore
       end
 
       def publish_event
-        Event.create!(event_id: SecureRandom.uuid, event_type: "Foo", data: {})
+        GlobalUpdater::Event.create!(event_id: SecureRandom.uuid, event_type: "Foo", data: {})
       end
 
       def global_status
-        ProjectionStatus.find_by(name: Consumer::GLOBAL_POSITION_NAME)
+        GlobalUpdater::ProjectionStatus.find_by(name: GlobalUpdater::GLOBAL_POSITION_NAME)
       end
 
       def publish_event_and_rollback
